requirements.txt
evdev==1.7.1
python-vlc==3.0.21203
